+++ platform-parse/src/main/java/net/dloud/platform/parse/kafka/listener/KafkaMessageListener.java
@@ -2,17 +2,13 @@ package net.dloud.platform.parse.kafka.listener;
 
 import lombok.extern.slf4j.Slf4j;
 import net.dloud.platform.common.domain.message.KafkaMessage;
-import net.dloud.platform.common.extend.StringUtil;
 import net.dloud.platform.extend.constant.PlatformConstants;
-import net.dloud.platform.parse.utils.SourceGet;
+import net.dloud.platform.parse.context.LocalContext;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.header.Header;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.listener.AcknowledgingMessageListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
-
-import static net.dloud.platform.parse.utils.SourceGet.kafkaProof;
 
 /**
  * @author QuDasheng
@@ -38,30 +34,27 @@ public final class KafkaMessageListener implements AcknowledgingMessageListener<
             return;
         }
 
-        final String proof;
-        final Iterable<Header> iterable = data.headers().headers(PlatformConstants.PROOF_KEY);
-        if (null != iterable && iterable.iterator().hasNext()) {
-            proof = new String(iterable.iterator().next().value());
-        } else {
-            proof = message.getProof();
+        final LocalContext local = LocalContext.fromKafka(data);
+        if (null == local) {
+            log.warn("[MESSAGE] 当前消息来源设置为空");
+            acknowledgment.acknowledge();
+            return;
         }
-        if (StringUtil.isBlank(kafkaProof.get()) || !proof.equalsIgnoreCase(kafkaProof.get())) {
-            kafkaProof.set(proof);
-        }
+        LocalContext.set(local);
 
         final String messageBean = message.getBean();
         if (null == messageBean) {
-            log.info("[MESSAGE] 当前消息消费方法为空, 不进行消费: proof={}", proof);
+            log.info("[MESSAGE] 当前消息消费方法为空, 不进行消费: 来源={}", local);
             acknowledgment.acknowledge();
             return;
         }
         if (null == message.getContent()) {
-            log.info("[MESSAGE] 当前消息内容为空, 不进行消费: proof={}", proof);
+            log.info("[MESSAGE] 当前消息内容为空, 不进行消费: 来源={}", local);
             acknowledgment.acknowledge();
             return;
         }
         if (message.getOnly() && !PlatformConstants.GROUP.equalsIgnoreCase(message.getGroup())) {
-            log.info("[MESSAGE] 当前消息设置为不消费: proof={}, {}", proof, message.getGroup());
+            log.info("[MESSAGE] 当前消息设置为不消费: 来源={}, {}", local, message.getGroup());
             acknowledgment.acknowledge();
             return;
         }
@@ -69,11 +62,13 @@ public final class KafkaMessageListener implements AcknowledgingMessageListener<
         try {
             // 处理接收到的消息
             messageExecutor.execute(message);
-            log.info("[MESSAGE] 消费完成开始提交offset: proof={}", proof);
+            log.info("[MESSAGE] 消费完成开始提交offset: 来源={}", local);
             acknowledgment.acknowledge();
-            log.info("[MESSAGE] 当前消息offset提交成功: proof={}", proof);
+            log.info("[MESSAGE] 当前消息offset提交成功: 来源={}", local);
         } catch (Exception e) {
             log.error("[MESSAGE] 消息消费过程中出现异常: ", e);
+        } finally {
+            LocalContext.remove();
         }
     }
 }