+++ platform-parse/src/main/java/net/dloud/platform/parse/module/MessageComponent.java
@@ -1,12 +1,12 @@
 package net.dloud.platform.parse.module;
 
-import com.alibaba.dubbo.rpc.RpcContext;
 import lombok.extern.slf4j.Slf4j;
 import net.dloud.platform.common.domain.message.KafkaMessage;
 import net.dloud.platform.extend.constant.CenterEnum;
 import net.dloud.platform.extend.constant.PlatformConstants;
 import net.dloud.platform.extend.constant.StartupConstants;
 import net.dloud.platform.extend.wrapper.AssertWrapper;
+import net.dloud.platform.parse.context.LocalContext;
 import net.dloud.platform.parse.kafka.SimpleMessage;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,10 +16,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.UUID;
 import java.util.stream.Collectors;
-
-import static net.dloud.platform.parse.utils.SourceGet.kafkaProof;
 
 /**
  * @author QuDasheng
@@ -31,6 +28,7 @@ import static net.dloud.platform.parse.utils.SourceGet.kafkaProof;
 public class MessageComponent {
     @Autowired
     private KafkaTemplate<Long, KafkaMessage> kafkaTemplate;
+
 
     /**
      * 发送消息到本系统
@@ -67,10 +65,11 @@ public class MessageComponent {
         AssertWrapper.notEmpty(topics, "要发送到的topic不能为空");
 
         //如何存在说明是外部调用以这个为准
-        final String proof = RpcContext.getContext().getAttachment(PlatformConstants.PROOF_KEY);
+        final LocalContext local = LocalContext.load();
         final KafkaMessage message = new KafkaMessage();
-        message.setProof(null == proof ? (null == kafkaProof.get()
-                ? UUID.randomUUID().toString() : kafkaProof.get()) : proof);
+
+        final String proof = local.getProof();
+        message.setProof(proof);
         message.setBean(bean);
         if (StartupConstants.RUN_MODE.equalsIgnoreCase(PlatformConstants.MODE_DEV)) {
             message.setOnly(true);
@@ -79,7 +78,7 @@ public class MessageComponent {
         }
         message.setGroup(PlatformConstants.GROUP);
         message.setContent(input);
-        log.info("[MESSAGE] proof={}, 发送到 {}, 目标 {}", message.getProof(), topics, bean);
+        log.info("[MESSAGE] 来源={}, 发送到 {}, 目标 {}", local, topics, bean);
 
         return kafkaTemplate.executeInTransaction(status -> {
             Long index = key;
@@ -89,18 +88,18 @@ public class MessageComponent {
 
             if (topics.size() > 1) {
                 for (String topic : topics) {
-                    kafkaTemplate.send(getRecord(topic, index, message));
+                    kafkaTemplate.send(getRecord(local, topic, index, message));
                 }
             } else {
-                kafkaTemplate.send(getRecord(topics.get(0), index, message));
+                kafkaTemplate.send(getRecord(local, topics.get(0), index, message));
             }
             return true;
         });
     }
 
-    private ProducerRecord<Long, KafkaMessage> getRecord(String topic, Long key, KafkaMessage message) {
+    private ProducerRecord<Long, KafkaMessage> getRecord(LocalContext local, String topic, Long key, KafkaMessage message) {
         final ProducerRecord<Long, KafkaMessage> record = new ProducerRecord<>(topic, key, message);
-        record.headers().add(PlatformConstants.PROOF_KEY, message.getProof().getBytes());
+        record.headers().add(PlatformConstants.LOCAL_KEY, local.toByte());
         return record;
     }
 }